# openapi.yaml
openapi: 3.0.3
info:
  title: WinOJ API
  description: |
    Windows 在线评测系统 API。支持多语言代码评测、安全沙箱执行、自定义计分脚本、比赛管理、文章系统、Rating 机制、AI 代码安全审查等功能。

    ## 认证
    大部分接口需携带 JWT Token：`Authorization: Bearer <access_token>`
    Access Token 有效期 15 分钟，过期后通过 `POST /auth/refresh` 刷新。

    ## 错误响应
    ```json
    { "code": 1|2|3|4|5|6, "reason": "ERR_xxx", "message": "..." }
    ```
    错误码：1=参数错误, 2=状态错误, 3=未找到, 4=限流, 5=未认证, 6=权限不足

    ## 角色层级
    user < teacher < admin < su
  version: 1.6.0
  contact:
    name: WinOJ
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
...
paths:
...
  /uploads:
    get:
      tags: [文件]
      summary: 文件列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 文件列表
    post:
      tags: [文件]
      summary: 上传文件
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: "最大 10MB"
      responses:
        '201':
          description: 上传成功
  /uploads/{filename}:
    get:
      tags: [文件]
      summary: 访问文件
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 文件内容
  /uploads/{id}:
    delete:
      tags: [文件]
      summary: 删除文件
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功

  # ===== DISCUSSIONS =====
...